requirements.txt
Flask
Flask-CORS
numpy
opencv-python
Pillow
